requirements.txt
requests
numpy
sounddevice
soundfile
pyperclip
keyboard
mutagen
customtkinter
packaging
pillow
black
isort
flake8